refactor(treemap): 兼容旧版矩阵树图下钻交互配置
